Change SkGlyph::fScalerContextBits to uint16_t.
This used to be uint8_t and was changed to uint32_t to soak up some
unused bits. However, it turns out that it was actually sitting in a
three byte region, so declaring this uint32_t accidentally made SkGlyph
larger. Changing this to uint16_t should allow it to soak up two of the
three unused bytes.

Bug: chromium:1315972
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/529812
Commit-Queue: Ben Wagner <bungeman@google.com>

+++ src/core/SkGlyph.h
@@ -483,7 +483,7 @@ private:
 
     // Used by the SkScalerContext to pass state from generateMetrics to generateImage.
     // Usually specifies which glyph representation was used to generate the metrics.
-    uint32_t  fScalerContextBits = 0;
+    uint16_t  fScalerContextBits = 0;
 
     // An SkGlyph can be created with just a packedID, but generally speaking some glyph factory
     // needs to actually fill out the glyph before it can be used as part of that system.
